requirements.txt
flask 
pandas
langchain 
langchain_community
# openai 
# selenium 
# playwright
python-dotenv
langchain-huggingface